+++ packages/blog/package.json
@@ -39,7 +39,7 @@
     "react-dom": "^16.4.1",
     "react-helmet": "^5.2.0",
     "react-icons": "^2.2.7",
-    "react-loadable": "^5.4.0",
+    "react-loadable": "5.5.0",
     "react-redux": "^5.0.6",
     "react-router-dom": "^4.2.2",
     "react-router-redux": "^4.0.8",
