+++ packages/blog/src/routes/Article/modules.js
@@ -47,3 +47,15 @@ export function getArticleDetail(id) {
         });
     };
 }
+
+export function fetchArticle() {
+    const {pathname} = document.location;
+    const id = pathname.substring(pathname.lastIndexOf('/') + 1);
+    return request.articleDetail(id).then(res => {
+        if (res.code === 0) {
+            return res.data;
+        } else {
+            return {};
+        }
+    });
+}
