+++ packages/blog/src/routes/Relative/stylesheets/relative.scss
@@ -1,9 +1,11 @@
 .relative-container {
   display: -ms-flexbox;
   display: flex;
+  -ms-flex-direction: column;
+  flex-direction: column;
   -ms-flex-align: stretch;
   align-items: stretch;
-  margin: 4.5rem auto;
+  margin: 0 auto;
   width: 100%;
 }
